LayerOptics Files now synchronized

+++ src/LayerOptics/tst/CircularPerforatedShade1.unit.cpp
@@ -1,5 +1,5 @@
 #include <memory>
-#include <gtest\gtest.h>
+#include <gtest/gtest.h>
 
 #include "UniformDiffuseBSDFLayer.hpp"
 #include "PerforatedCell.hpp"
@@ -118,12 +118,12 @@ TEST_F( TestCircularPerforatedShade1, TestSolarProperties )
   correctResults.push_back( 0.063662 );
 
   vector< double > calculatedResults;
-  for( int i = 0; i < size; ++i ) {
+  for( size_t i = 0; i < size; ++i ) {
     calculatedResults.push_back( (*aT)[i][i] );
   }
 
   EXPECT_EQ( correctResults.size(), calculatedResults.size() );
-  for( int i = 0; i < size; ++i ) {
+  for( size_t i = 0; i < size; ++i ) {
     EXPECT_NEAR( correctResults[i], calculatedResults[i], 1e-5 );
   }
 
@@ -172,13 +172,13 @@ TEST_F( TestCircularPerforatedShade1, TestSolarProperties )
   correctResults.push_back( 0.063662 );
 
   calculatedResults.clear();
-  for( int i = 0; i < size; ++i ) {
+  for( size_t i = 0; i < size; ++i ) {
     //cout << (*aT)[ 0 ][ i ] << endl;
     calculatedResults.push_back( (*aT)[ 0 ][ i ] );
   }
 
   EXPECT_EQ( correctResults.size(), calculatedResults.size() );
-  for( int i = 0; i < size; ++i ) {
+  for( size_t i = 0; i < size; ++i ) {
     EXPECT_NEAR( correctResults[i], calculatedResults[i], 1e-5 );
   }
 
@@ -229,12 +229,12 @@ TEST_F( TestCircularPerforatedShade1, TestSolarProperties )
   correctResults.push_back( 0.238732 );
 
   calculatedResults.clear();
-  for( int i = 0; i < size; ++i ) {
+  for( size_t i = 0; i < size; ++i ) {
     calculatedResults.push_back( (*aRf)[ 0 ][ i ] );
   }
 
   EXPECT_EQ( correctResults.size(), calculatedResults.size() );
-  for( int i = 0; i < size; ++i ) {
+  for( size_t i = 0; i < size; ++i ) {
     EXPECT_NEAR( correctResults[i], calculatedResults[i], 1e-5 );
   }
 
@@ -285,12 +285,12 @@ TEST_F( TestCircularPerforatedShade1, TestSolarProperties )
   correctResults.push_back( 0.210085 );
 
   calculatedResults.clear();
-  for( int i = 0; i < size; ++i ) {
+  for( size_t i = 0; i < size; ++i ) {
     calculatedResults.push_back( (*aRb)[ 0 ][ i ] );
   }
 
   EXPECT_EQ( correctResults.size(), calculatedResults.size() );
-  for( int i = 0; i < size; ++i ) {
+  for( size_t i = 0; i < size; ++i ) {
     EXPECT_NEAR( correctResults[i], calculatedResults[i], 1e-5 );
   }
